requirements.txt
asarPy==1.0.1
asgiref==3.8.1
attrs==23.2.0
coverage==7.5.4
Django==5.0.6
django-rest-framework==0.1.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.8.0
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.5.0
uritemplate==4.1.1
